+++ acl-engine/src/main/java/javasabr/mqtt/acl/engine/builder/RuleContainerBuilder.java
@@ -1,4 +1,4 @@
-package javasabr.mqtt.service.acl.builder;
+package javasabr.mqtt.acl.engine.builder;
 
 import java.util.Collections;
 import java.util.EnumMap;
