+++ acl-engine/src/main/java/javasabr/mqtt/acl/engine/builder/TopicMatcherBuilder.java
@@ -1,5 +1,5 @@
 //file:noinspection unused
-package javasabr.mqtt.service.acl.builder;
+package javasabr.mqtt.acl.engine.builder;
 
 import javasabr.mqtt.acl.engine.model.matcher.TopicFilterMatcher;
 import javasabr.mqtt.acl.engine.model.matcher.TopicNameMatcher;
