+++ acl-engine/src/main/java/javasabr/mqtt/acl/engine/model/rule/Rule.java
@@ -1,7 +1,7 @@
 package javasabr.mqtt.acl.engine.model.rule;
 
-import javasabr.mqtt.model.MqttUser;
 import javasabr.mqtt.acl.engine.model.Action;
+import javasabr.mqtt.model.MqttUser;
 import javasabr.mqtt.model.acl.Operation;
 import javasabr.mqtt.model.topic.AbstractTopic;
